app/requirements.txt
pandas
Flask
Flask-Cors
flask-restx
flask-sqlalchemy
mysql-connector-python